% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_penepma_z_shifts.R
\name{calculate_penepma_z_shifts}
\alias{calculate_penepma_z_shifts}
\title{Calculate a list of penepma z-shifts given a list of layer spacings}
\usage{
calculate_penepma_z_shifts(l_nm_shifts)
}
\arguments{
\item{l_nm_shifts}{A list of layer thicknesses in nm.
Example c(20, 50, 100, 1.0e+07)}
}
\value{
l_cum_sum A list of the cumulative sum values. These are
                  printed with 15 digits to match penepma's output
}
\description{
Calculate a list of penepma z-shifts given a list of layer spacings
}
\examples{
library(rpenepma)
# data
l_nm_shifts <- c(20, 50, 100, 1.0e+07)
l_cum_shifts_cm <- calculate_penepma_z_shifts(l_nm_shifts)
print(l_cum_shifts_cm)
}
