% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_to_nm.R
\name{cm_to_nm}
\alias{cm_to_nm}
\title{Convert cm to nm for processing PENEPMA data}
\usage{
cm_to_nm(cm)
}
\arguments{
\item{nm}{A number. The value to convert.}
}
\value{
cm The result in cm
}
\description{
Convert cm to nm for processing PENEPMA data
}
\examples{
library(rpenepma)
print(cm_to_nm(20.0))
}
