% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_densities_data.R
\docType{data}
\name{elemental_densities}
\alias{elemental_densities}
\title{Densities for the elements.}
\format{A dataframe/tibble with 96 observations of the variables
 \code{Density}, \code{Unit}, \code{Name}, \code{Symbol} and \code{Z}}
\source{
https://www.angstromsciences.com/density-elements-chart
}
\usage{
data(elemental_densities)
}
\description{
A dataframe/tibble containing the Density, Unit,
 Name, Symbol, and Z for 96 element.
}
\author{
angstromsciences
}
