% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_simulation_finish_time.R
\name{estimate_simulation_finish_time}
\alias{estimate_simulation_finish_time}
\title{Estimate the finish time for a simulation, given a prior estimate}
\usage{
estimate_simulation_finish_time(start_date, start_time, base_hrs, base_min,
  base_sec, multiplier)
}
\arguments{
\item{start_date}{A string with the starting date.
Example: "2018-07-27"}

\item{start_time}{A string with the starting time.
Example: "15:01:32"}

\item{base_hrs}{A number. Number of hrs for base simulation.
Example: 0}

\item{base_min}{A number. Number of min for base simulation.
Example: 10}

\item{base_sec}{A number. Number of sec for base simulation.
Example: 15}

\item{multiplier}{A number. The ratio based on parameterers.
Example: For 5X the trajectories use 5.0.}
}
\value{
None. Prints estimated finish time
}
\description{
Estimate the finish time for a simulation, given a prior estimate
}
\examples{
library(rpenepma)
estimate_simulation_finish_time("2018-07-27", "15:01:32",
                                0, 10, 15, 5.0)

}
