% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_atomic_symbol_from_z.R
\name{get_atomic_symbol_from_z}
\alias{get_atomic_symbol_from_z}
\title{Get the atomic symbol from Z}
\usage{
get_atomic_symbol_from_z(z)
}
\arguments{
\item{z}{The atomic number for an element}
}
\value{
A string. Atomic symbol
}
\description{
Get the atomic symbol from Z
}
