% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_penepma_spectra_info.R
\name{get_penepma_spectra_info}
\alias{get_penepma_spectra_info}
\title{Retrieve the spectrum information from a PENEPMA spectrum}
\usage{
get_penepma_spectra_info(fi)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}
}
\value{
a tibble with the e_start_eV, e_end_eV, n_channels and ev_per_ch
}
\description{
Retrieve the spectrum information from a PENEPMA spectrum
}
