% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_to_cm.R
\name{nm_to_cm}
\alias{nm_to_cm}
\title{Convert nm to cm for processing PENEPMA data}
\usage{
nm_to_cm(nm)
}
\arguments{
\item{nm}{A number. The value to convert.}
}
\value{
cm The result in cm
}
\description{
Convert nm to cm for processing PENEPMA data
}
\examples{
library(rpenepma)
print(cm_to_nm(2.0e-6))
}
