% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma12_get_number_of_showers.R
\name{penepma12_get_number_of_showers}
\alias{penepma12_get_number_of_showers}
\title{Retrieve the number of showers from a penepma12 simulation}
\usage{
penepma12_get_number_of_showers(fi)
}
\arguments{
\item{fi}{String. The path to the penepma.dat file. ex: "./penepma.dat"}
}
\value{
the number of "showers" i.e. trajectories computed.  (numeric)
}
\description{
Penepma12 adds to the penepam.dat file. There is no penepma-res.dat file.
like penepma16 produces. This function extracts the latest count of the
number of showers.
}
