% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma16_read_map_data.R
\name{penepma16_read_map_data}
\alias{penepma16_read_map_data}
\title{penepma16_read_map_data - read the pe-map-0#-depth.dat file
and return a tibble with data values}
\usage{
penepma16_read_map_data(fi)
}
\arguments{
\item{fi}{A string. the path to the desired file, ex: './pe-map-01-depth.dat'}
}
\value{
A tibble with depth [nm], dens.mu (mean) and dens.unc (std.error)
}
\description{
A helper function for processing map data
}
