% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma16_read_map_hdr.R
\name{penepma16_read_map_hdr}
\alias{penepma16_read_map_hdr}
\title{penepma16_read_map_hdr - read the pe-map-0#-depth.dat file
and return a list with Z and the transition}
\usage{
penepma16_read_map_hdr(fi)
}
\arguments{
\item{fi}{A string. the path to the desired file, ex: './pe-map-01-depth.dat'}
}
\value{
A list. Element 1 is the Z value and element 2 is the trasnition
}
\description{
A helper function for processing depth data
}
