% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_calc_k_ratio.R
\name{penepma_calc_k_ratio}
\alias{penepma_calc_k_ratio}
\title{Calculate the k-ratio from penepma intensities}
\usage{
penepma_calc_k_ratio(str_samp, str_kr, tib_unk, tib_std, z, s0, s1,
  n_rnd = 5)
}
\arguments{
\item{str_samp}{A string to describe the sample. Ex: "200 nm C on Fe"}

\item{str_kr}{A string to describe the K-ratio Ex: "C Ka"}

\item{tib_unk}{A tibble containing the penepma intensity data for
the unknown.}

\item{tib_std}{A tibble comtaining the penepma intensity data for
the standard.}

\item{z}{The atomic number for the element for the K-ratio.
Ex: 6 for C.}

\item{s0}{The number for the originating shell from Table 7.2.
Ex: 1 for K.}

\item{s1}{The number for the filling shell from Table 7.2.
Ex: 4 for L3.}

\item{n_rnd}{The number of digits to round values. Default: 5.}
}
\value{
A tibble containing the kratio and the uncertainty
}
\description{
Calculate the k-ratio for a desired transition from R tibbles that
contain intensity data. Also compute the uncertainty and relative
uncertainty.
}
