% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_copy_list_of_files.R
\name{penepma_copy_list_of_files}
\alias{penepma_copy_list_of_files}
\title{Copy a list of files from simulation to analysis directory}
\usage{
penepma_copy_list_of_files(in_dir, out_dir, fi_spec, verbose = FALSE)
}
\arguments{
\item{in_dir}{String. The path to the simulation dir. ex: "~Desktop/sim"}

\item{out_dir}{String. The path to the anaysis dir. ex: "~Desktop/ana"}

\item{fi_spec}{String. A file specification. ex: "*-01.dat"}

\item{verbose}{Boolean, Default: FALSE. If TRUE print list of files.}
}
\value{
none.
}
\description{
PENEPMA generates many output data files durming a simulation. Often one
wants to analyze "on-the-fly". This function helps copy them to an anaysis
directory
}
