% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_add_body.R
\name{penepma_geo_add_body}
\alias{penepma_geo_add_body}
\title{Add a body}
\usage{
penepma_geo_add_body(indx, str_cmt)
}
\arguments{
\item{indx}{integer. Body number. Corresponde to material.}

\item{str_cmt}{string. A comment. Ex: top layer}
}
\value{
none. Exports by cat to  console or file as set previously
}
\description{
A helper function to create .geo files
}
