% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_add_surf_cyl.R
\name{penepma_geo_add_surf_cyl}
\alias{penepma_geo_add_surf_cyl}
\title{Add a cylindrical surface}
\usage{
penepma_geo_add_surf_cyl(indx, str_cmt, rad_cm)
}
\arguments{
\item{indx}{integer. Surface number. Starts with 1.}

\item{str_cmt}{string. A comment. Ex:  Cylinder, 1 cm radius}

\item{rad_cm}{The radius in cm Ex: 1.0.}
}
\value{
none. Exports by cat to  console or file as set previously
}
\description{
A helper function to create penepma .geo files
}
