% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_add_surf_z.R
\name{penepma_geo_add_surf_z}
\alias{penepma_geo_add_surf_z}
\title{Add a Z surface}
\usage{
penepma_geo_add_surf_z(indx, str_cmt, l_index, cum_z_shift_nm)
}
\arguments{
\item{indx}{integer. Surface number. Starts with 1.}

\item{str_cmt}{string. A comment. Ex:  Plane Z=0 - surface plane}

\item{l_index}{list of 5 integer indices. Ex: c(0, 0, 0, 1, 0)}

\item{cum_z_shift_nm}{The CUMULATIVE z-shift of the plane in nm. Converted
to cm internally. These have negative values. Default is NA which is used
for the top surface that does not have a Z-shift.}
}
\value{
none. Exports by cat to console or file as set previously
}
\description{
A helper function to create .geo files
}
