% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_fini.R
\name{penepma_geo_fini}
\alias{penepma_geo_fini}
\title{penepma_geo_finish the geo file}
\usage{
penepma_geo_fini()
}
\value{
none. Exports by cat to console or file as set previously
}
\description{
A helper script to create .geo files.
}
