% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_hdr_0.R
\name{penepma_geo_hdr_0}
\alias{penepma_geo_hdr_0}
\title{Create a header line of repeated zeros.}
\usage{
penepma_geo_hdr_0()
}
\value{
none. Exports by cat to  console or file as set previously
}
\description{
A helper function to create .geo files
}
