% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_hdr_c.R
\name{penepma_geo_hdr_c}
\alias{penepma_geo_hdr_c}
\title{Create a header line with just 'C'}
\usage{
penepma_geo_hdr_c()
}
\value{
none. Exports by cat to console or file as set previously
}
\description{
A helper function to create .geo files
}
