% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_hdr_cmt.R
\name{penepma_geo_hdr_cmt}
\alias{penepma_geo_hdr_cmt}
\title{Create a header line with a comment}
\usage{
penepma_geo_hdr_cmt(str_cmt)
}
\arguments{
\item{str_cmt}{string. Example: "Cylindrical layers - distances all in cm"}
}
\value{
none. Exports by cat to  console or file as set previously
}
\description{
A helper function to create .geo files
}
