% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_geo_hdr_x.R
\name{penepma_geo_hdr_x}
\alias{penepma_geo_hdr_x}
\title{Create a header line of repeated x}
\usage{
penepma_geo_hdr_x()
}
\value{
none. Exports by cat to  console or file as set previously
}
\description{
A helper function to generate .geo files
}
