% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_ev_per_ch.R
\name{penepma_get_ev_per_ch}
\alias{penepma_get_ev_per_ch}
\title{Retrieve the ev/ch from a PENEPMA spectrum}
\usage{
penepma_get_ev_per_ch(fi)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}
}
\value{
the ev/channel (numeric)
}
\description{
Retrieve the ev/ch from a PENEPMA spectrum
}
