% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_intensities.R
\name{penepma_get_intensities}
\alias{penepma_get_intensities}
\title{Retrieve the intensities from a PENEPMA spectrum}
\usage{
penepma_get_intensities(fi)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}
}
\value{
the intensities (a tibble)
}
\description{
Retrieve the intensities from a PENEPMA spectrum
}
