% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_intensity_z_s0_s1.R
\name{penepma_get_intensity_z_s0_s1}
\alias{penepma_get_intensity_z_s0_s1}
\title{Retrieve the intensity and uncertainty for a specified transition for an
element and calculate the SNR. Adds the atomic symbol to the return.}
\usage{
penepma_get_intensity_z_s0_s1(tib, z, s0, s1)
}
\arguments{
\item{tib}{A tibble. The tibble with the intensities from a penepma spectrum}

\item{z}{The atomic number for an element}

\item{s0}{The number for the originating shell from Table 7.2. Ex: 1 for K.}

\item{s1}{The number for the filling shel lfrom Table 7.2. Ex: 4 for L3.}
}
\value{
A tibble. The Symbol, atomic number, S0, S1, mean intensity,
        mean intensity std error,and mean intensity SNR.
}
\description{
Retrieve the intensity and uncertainty for a specified transition for an
element and calculate the SNR. Adds the atomic symbol to the return.
}
