% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_max_primary_intensity_z.R
\name{penepma_get_max_primary_intensity_z}
\alias{penepma_get_max_primary_intensity_z}
\title{Retrieve the maximum intensity for an element}
\usage{
penepma_get_max_primary_intensity_z(tib, z)
}
\arguments{
\item{tib}{A tibble. The tibble with the intensities from a penepma spectrum}

\item{z}{The atomic number for an element}
}
\value{
the intensities transition with the maximum primary intensity (a tibble)
}
\description{
Retrieve the maximum intensity for an element
}
