% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_max_total_intensity_z.R
\name{penepma_get_max_total_intensity_z}
\alias{penepma_get_max_total_intensity_z}
\title{Retrieve the maximum total intensity qnd uncertqainty for an element}
\usage{
penepma_get_max_total_intensity_z(tib, z)
}
\arguments{
\item{tib}{A tibble. The tibble with the intensities from a penepma spectrum}

\item{z}{The atomic number for an element}
}
\value{
the transition and the maximum total intensity and uncertainty (a tibble)
}
\description{
Retrieve the maximum total intensity qnd uncertqainty for an element
}
