% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_number_of_showers.R
\name{penepma_get_number_of_showers}
\alias{penepma_get_number_of_showers}
\title{Retrieve the number of showers from a penepma simulation}
\usage{
penepma_get_number_of_showers(fi)
}
\arguments{
\item{fi}{String. The path to the penepma-res.dat file. ex: "./penepma-res.dat"}
}
\value{
the number of "showers" i.e. trajectories computed.  (numeric)
}
\description{
Retrieve the number of showers from a penepma simulation
}
