% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_prz_data.R
\name{penepma_get_prz_data}
\alias{penepma_get_prz_data}
\title{Retrieve the phi rho z data from a line in a penepma simulation}
\usage{
penepma_get_prz_data(fi)
}
\arguments{
\item{fi}{String. The path to the pe-prz-140104-01.dat file.
ex: "./ pe-prz-140104-0.dat"}
}
\value{
A tibble with the prz data. There is a lot of data for post-processing
}
\description{
Read the data from the pe-prz-140104-01 file . In this case 14 ia the atomic
number (Si) 01 indicates a K shell and 04 inficates an L3 shell. The 01 
signifies detector 1.
}
