% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_measure_simulation_time.R
\name{penepma_measure_simulation_time}
\alias{penepma_measure_simulation_time}
\title{Measure the simulation time of a penepma simulation}
\usage{
penepma_measure_simulation_time(sim_dir)
}
\arguments{
\item{sim_dir}{String. The path to the directory with the penepma simulation
files. Example: "C:/sim/penepma16/20-nm-C-250-nm-ZnO-Silica".}
}
\value{
the current simulation time. An integer of class "POSIXct"
}
\description{
Measure the time difference between the creation time of the pe-material.dat
(not changed after the simulation starts) and the modification time of the
pe-spect-01.dat which is modified at the end of every sump time. We do this
in the simulation directory to preserve the file time stamps.
}
