% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_plot_spectrum.R
\name{penepma_plot_spectrum}
\alias{penepma_plot_spectrum}
\title{Plot a PENEPMA spectrum file}
\usage{
penepma_plot_spectrum(fi, title)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}

\item{title}{String. The title for the plot.}
}
\value{
a ggplot
}
\description{
Plot a PENEPMA spectrum file
}
