% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_plot_spectrum_logy.R
\name{penepma_plot_spectrum_logy}
\alias{penepma_plot_spectrum_logy}
\title{Plot a PENEPMA spectrum file with a log Y scale}
\usage{
penepma_plot_spectrum_logy(fi, title)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}

\item{title}{String. The title for the plot.}
}
\value{
a ggplot
}
\description{
Plot a PENEPMA spectrum file with a log Y scale
}
