% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_read_raw_data.R
\name{penepma_read_raw_data}
\alias{penepma_read_raw_data}
\title{Retrieve the raw channel data from a PENEPMA spectrum}
\usage{
penepma_read_raw_data(fi, min_intensity_clip = 1e-12,
  int_scale_factor = 1e+09)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}

\item{min_intensity_clip}{numeric. The minimum value to clip the intensity. Default: 1.0e-12.}

\item{int_scale_factor}{numeric. Intensity scale factor. Default: 1.0e9.}
}
\value{
a tibble
}
\description{
Retrieve the raw channel data from a PENEPMA spectrum
}
