% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_report_rel_unc.R
\name{penepma_report_rel_unc}
\alias{penepma_report_rel_unc}
\title{A tidy retrieval of the relative uncertainty of a transition from a
tibble/dataframe}
\usage{
penepma_report_rel_unc(tib_in, z_val, val_s0, val_s1)
}
\arguments{
\item{tib_in}{A tibble. This contains the intensity data. Usually created
with penepma_get_intensities()}

\item{z_val}{A number. The atomic number of the desired element. Ex: 29 (Cu)}

\item{val_s0}{A number. The penepma value from Table 7.2 for the first shell
of the transition. Ex: K = 1, L3 = 5, M5 = 9.}

\item{val_s1}{A number. The penepma value from Table 7.2 for the second shell
of the transition. (see example above or table 7.3)}
}
\value{
a tibble with a single row containing the relative uncertainty.
}
\description{
A tidy retrieval of the relative uncertainty of a transition from a
tibble/dataframe
}
