% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_to_msa.R
\name{penepma_to_msa}
\alias{penepma_to_msa}
\title{Convert a PENEPMA spectrum file to MSA format}
\usage{
penepma_to_msa(datFile, msaFile, e0, title, scale_fact = 1,
  owner = "Penepma", bDebug = FALSE)
}
\arguments{
\item{datFile}{string - the path to the 'pe-spect-##.dat'
file}

\item{msaFile}{string - path fpr the output 'my-spec.msa' file.}

\item{e0}{number - the voltage in kV for the simulation}

\item{title}{string - label for the spectrum (default "Penepma")}

\item{scale_fact}{number - number of counts to set for minimum value.
Adjust if needed to be sure the Duane--Hunt limit is displayed.
Default = 1.0.}

\item{owner}{string - label for the scientist}

\item{bDebug}{boolean - default FALSE. If TRUE, it prints headers
for diagnostics. Otherwise the function is silent.}
}
\value{
None

It silently writes the .msa file
}
\description{
Convert a spectrum simulated by PENEPMA into MSA format to
be read into DTSA-II (where we have markers...) The data is
scaled for DTSA-II.
}
\examples{
# Not run
# inF <- system.file("extdata", "pe-spect-01.dat", package = "rEDS")
# ouF <- './penepma/k411-15kV.msa'
# penepma_to_msa(inF, ouF, 15, 'simulation')

}
\keyword{keywords}
