% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_penepma_res.R
\name{poll_penepma_res}
\alias{poll_penepma_res}
\title{Periodically poll the penepma-res.dat file and generate a tibble
containing the simulation time (in sec), the number of trajctories,
and the relative uncertainty. This is a wrapper for the function
proc_penepma16_res.}
\usage{
poll_penepma_res(sim_dir, sleep_sec = 120, max_iter = 20000,
  skip_time = 6, skip_traj = 2, skip_unc = 33, min_lines = 10,
  clean = FALSE, verbose = FALSE)
}
\arguments{
\item{sim_dir}{The path to the directory with penepma.res file.}

\item{sleep_sec}{An integer. The number of sec between reading the
file. Set to the dump time for the simulation}

\item{max_iter}{An integer. The maximum number of iterations. Default
is 20000.}

\item{skip_time}{An integer. The number of lines to skip to read the
line containing the simulation time. Default:6.}

\item{skip_traj}{An integer. The number of lines to skip to read the
line containing the number of trajectories, Default: 2 (additional
lines).}

\item{skip_unc}{An integer. The number of lines to skip to read the
line with the uncertainty value. Default: 33 (additional lines). This
line changes frequently because of the change in number of detectors.}

\item{min_lines}{An integer. The minimum lines the file should have
to process. Helps if the file polled while penepma was writing. 
Default: 10.}

\item{clean}{A boolean. The underlying proc_penepma16_res copies the
penepma-res.dat file to work.dat. If TRUE, this file is deleted.
Default is FALSE.}

\item{verbose}{A boolean. Write messages when TRUE. Default is FALSE.}
}
\description{
Periodically poll the penepma-res.dat file and generate a tibble
containing the simulation time (in sec), the number of trajctories,
and the relative uncertainty. This is a wrapper for the function
proc_penepma16_res.
}
