% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend_sample_id_max_int_tib.R
\name{prepend_sample_id_max_int_tib}
\alias{prepend_sample_id_max_int_tib}
\title{Prepend a sample ID to a maximum intensity tibble}
\usage{
prepend_sample_id_max_int_tib(tib, sample_id)
}
\arguments{
\item{tib}{An intensity tibble generated by 
penepma_get_max_total_intensity_z or
penepma_get_max_primary_intensity_z.}

\item{sample_id}{String. A unique sample identifer}
}
\value{
a tibble. The sample_ID will be the first column
followed by Symbol, IZ, S0, S1, Int.mu, Int.se, Int.snr
}
\description{
Prepend a sample ID to a maximum intensity tibble
}
