% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_penepma16_res.R
\name{proc_penepma16_res}
\alias{proc_penepma16_res}
\title{Get the time, number of trajectories, and relative uncertainty from a
penepma16 simulation. Sometimes the lines in the file are shifted.}
\usage{
proc_penepma16_res(sim_dir, skip_time = 6, skip_traj = 2,
  skip_unc = 33, min_lines = 10, clean = FALSE)
}
\arguments{
\item{sim_dir}{The path to the directory with penepma.res file.}

\item{skip_time}{An integer. The number of lines to skip to read the
line containing the simulation time. Default:6.}

\item{skip_traj}{An integer. The number of lines to skip to read the
line containing the number of trajectories, Default: 2
(additional lines).}

\item{skip_unc}{An integer. The number of lines to skip to read the
line with the uncertainty value. Default: 33 (additional lines).}

\item{min_lines}{An integer. Line limit to be considered a bad read.
fail gracefully, returning NAs that we can remove later.}

\item{clean}{A boolean. If TRUE, delete the work.dat file.
Default: FALSE}
}
\value{
a tibble with the simulation time (sec), number of
trajectories, and the relative uncertainty of the REFLIN transition
}
\description{
Get the time, number of trajectories, and relative uncertainty from a
penepma16 simulation. Sometimes the lines in the file are shifted.
}
