% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prz_wrapper.R
\name{prz_wrapper}
\alias{prz_wrapper}
\title{Get the data we need to plot a prz curve}
\usage{
prz_wrapper(sim_dir, prz_fi_name)
}
\arguments{
\item{sim_dir}{String. The directory containing the prz file.}

\item{prz_fi_name}{String. The name of the prz file WITHOUT the .dat ext.
ex: "pe-prz-140104-01"}
}
\value{
A tibble with the prz data. This can be plotted using R. The function
will also write a file to the simulation directory suitable for processing
with gnuplot for a quick graph.
}
\description{
Use penepma_get_prz_data() to get the data, make the data tidy by converting
distances to nm and extracting just the data we need.
}
