% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pypenelope_get_number_of_showers.R
\name{pypenelope_get_number_of_showers}
\alias{pypenelope_get_number_of_showers}
\title{Retrieve the number of showers from a pypenelope simulation}
\usage{
pypenelope_get_number_of_showers(fi)
}
\arguments{
\item{fi}{String. The path to the penepma.log file. ex: "./penepma.log"}
}
\value{
the number of "showers" i.e. trajectories computed.  (numeric)
}
\description{
Pypenelope seems to write the nnumber of showers as the first entry in the
penepma.log file.
}
