% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_to_hr.R
\name{sec_to_hr}
\alias{sec_to_hr}
\title{A convenience funtion to convert seconds to hrs}
\usage{
sec_to_hr(sec, digits = 3)
}
\arguments{
\item{sec}{A number. The number of seconds}

\item{digits}{An integer. Default 3. Number of digits of precision. Default: 3.}
}
\value{
A number. The number of hrs.
}
\description{
Penepma reports simulation times in seconds. This converts to hrs.
}
