% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_to_min.R
\name{sec_to_min}
\alias{sec_to_min}
\title{A convenience funtion to convert seconds to minutes}
\usage{
sec_to_min(sec, digits = 2)
}
\arguments{
\item{sec}{A number. The number of seconds}

\item{digits}{An integer. Default 2. Number of digits of precision. Default: 2.}
}
\value{
A number. The number of minutes.
}
\description{
Penepma reports simulation times in seconds. This converts to minutes.
}
