% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_penepma16_res.R
\name{test_penepma16_res}
\alias{test_penepma16_res}
\title{Find the lines containing simulation time, number of trajectories,
and relative uncertainty from a penepma16 simulation. The lines in
the file are shifted, typically because of changes in the number of
detectors. Use this to find offsets for the function
proc_penepma16_res}
\usage{
test_penepma16_res(sim_dir, skip_time = 6, skip_traj = 2,
  skip_unc = 33)
}
\arguments{
\item{sim_dir}{The path to the directory with penepma.res file.}

\item{skip_time}{An integer. The number of lines to skip to read the line
containing the simulation time. Default:6.}

\item{skip_traj}{An integer. The number of lines to skip to read the line
containing the number of trajectories, Default: 2 (additional lines).}

\item{skip_unc}{An integer. The number of lines to skip to read the line
with the uncertainty value. Default: 33 (additional lines).}
}
\value{
None. Prints lines chosen.
}
\description{
Find the lines containing simulation time, number of trajectories,
and relative uncertainty from a penepma16 simulation. The lines in
the file are shifted, typically because of changes in the number of
detectors. Use this to find offsets for the function
proc_penepma16_res
}
