% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma16_get_time_traj_uncert.R
\name{penepma16_get_time_traj_uncert}
\alias{penepma16_get_time_traj_uncert}
\title{Get the time, number of trajectories, and relative uncertainty from a
penepma16 simulation}
\usage{
penepma16_get_time_traj_uncert(sim_dir)
}
\arguments{
\item{sim_dir}{The path to the directory with penepma.res file.}
}
\value{
a tibble with the simulation time (sec), number of
trajectories, and the relative uncertainty of the REFLIN transition
}
\description{
Get the time, number of trajectories, and relative uncertainty from a
penepma16 simulation
}
