% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_get_max_primary_intensity_z.R
\name{penepma_get_max_primary_intensity_z}
\alias{penepma_get_max_primary_intensity_z}
\title{Retrieve the maximum primary intensity and uncertainty for an element
and calculate the SNR}
\usage{
penepma_get_max_primary_intensity_z(tib, z)
}
\arguments{
\item{tib}{A tibble. The tibble with the intensities from a penepma spectrum}

\item{z}{The atomic number for an element}
}
\value{
A tibble. The transition, maximum primary intensity, uncertainty, and SNR.
}
\description{
Retrieve the maximum primary intensity and uncertainty for an element
and calculate the SNR
}
