% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penepma_read_raw_data.R
\name{penepma_read_raw_data}
\alias{penepma_read_raw_data}
\title{Retrieve the raw channel data from a PENEPMA spectrum}
\usage{
penepma_read_raw_data(fi)
}
\arguments{
\item{fi}{String. The path to the file. ex: "./pe-spect-01.dat"}
}
\value{
a data frame
}
\description{
Retrieve the raw channel data from a PENEPMA spectrum
}
