% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pypenelope_get_number_of_showers.R
\name{pypenelope_get_number_of_showers}
\alias{pypenelope_get_number_of_showers}
\title{Retrieve the number of showers from a pypenelope simulation}
\usage{
pypenelope_get_number_of_showers(dir)
}
\arguments{
\item{dir}{String. The directory with penepma.log file. ex: "here()"}
}
\value{
the number of "showers" i.e. trajectories computed. (numeric)
}
\description{
If pypenelope ends on a time or uncertainty measure, the number of
trajectories seems to be written in the penepma.log file. If the
simulation is time limited, the data must be retrieved from the
penepma.dat file.
}
